refactor encapsulation for likes
